src/cs-agent/requirements.txt
google-adk[a2a]
uvicorn>=0.25.0
fastapi>=0.115.0
requests>=2.31.0
PyJWT>=2.8.0